+++ style/index.js
@@ -1,0 +1,17 @@
+import '../src/Table.less';
+import 'kuma-base/core.less';
+import 'uxcore-animate/style';
+import 'uxcore-button/style';
+import 'uxcore-button-group/style';
+import 'uxcore-checkbox-group/style';
+import 'uxcore-collapse/style';
+import 'uxcore-dropdown/style';
+import 'uxcore-empty-data/style';
+import 'uxcore-icon/style';
+import 'uxcore-menu/style';
+import 'uxcore-pagination/style';
+import 'uxcore-popover/style';
+import 'uxcore-radiogroup/style';
+import 'uxcore-select2/style';
+import 'uxcore-tooltip/style';
+import 'uxcore-tree/style';